Feat/post

See merge request react-node-track/class01/web-project-final/team02/dogaja-frontend!101

+++ src/app/components/AddressBookModal.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client";
 
 import React, { useState, useEffect } from "react";
 import Modal from "./Modal";
@@ -9,9 +9,10 @@ import FriendAddModal from "./FriendAddModal";
 import { API } from "@/app/utils/api";
 import { mainApi } from "@/app/utils/mainApi";
 import { SuccessAlert, ErrorAlert } from "@/app/utils/toastAlert";
+import { IoPersonCircleOutline } from "react-icons/io5";
 
 interface Friend {
-  id: number;
+  id: string;
   name: string;
 }
 
@@ -27,16 +28,18 @@ const AddressBookModal: React.FC<{
   useEffect(() => {
     getFriendList();
   }, []);
-  
+
   const getFriendList = async () => {
     const { FRIENDS_LIST_GET } = API.FRIENDS;
     try {
       const res = await mainApi({
         url: FRIENDS_LIST_GET,
         method: "GET",
+        withAuth: true,
       });
       if (res.status === 200) {
-        setFriends(res.data.friends || []);
+        const data = (res.data as Friend[]) || [];
+        setFriends(data);
       }
     } catch (e) {
       console.error(e);
@@ -44,15 +47,22 @@ const AddressBookModal: React.FC<{
     }
   };
 
-  const handleDelete = async (id: number) => {
+  const handleDelete = async (id: string) => {
+    const deleteConfirm = confirm("정말 친구 목록에서 삭제하시겠습니까?");
+    if (!deleteConfirm) {
+      return;
+    }
     const { FRIENDS_DELETE } = API.FRIENDS;
     try {
       const res = await mainApi({
-        url: FRIENDS_DELETE,
+        url: FRIENDS_DELETE(id),
         method: "DELETE",
+        withAuth: true,
       });
       if (res.status === 200) {
-        setFriends((prevFriends) => prevFriends.filter((friend) => friend.id !== id));
+        setFriends((prevFriends) =>
+          prevFriends.filter((friend) => friend.id !== id),
+        );
         SuccessAlert("친구 목록에서 삭제되었습니다.");
       }
     } catch (e) {
@@ -73,15 +83,15 @@ const AddressBookModal: React.FC<{
   return (
     <>
       <Modal isOpen={isOpen} onClose={onClose} title="주소록">
-        <div className="max-h-60 overflow-y-auto mb-4">
+        <div className="min-h-60 max-h-60 overflow-y-auto mb-4">
           <ul className="mb-4">
             {friends.map((friend) => (
               <li
                 key={friend.id}
-                className="flex justify-between items-center py-2 border-b border-gray-300"
+                className="flex justify-between items-center py-2 px-2 border-b border-gray-300"
               >
-                <span className="flex items-center">
-                  <div className="mr-3 py-3 px-1 border rounded-lg bg-gray-300">이미지</div>
+                <span className="flex items-center gap-2">
+                  <IoPersonCircleOutline className="text-gray-300 w-[50px] h-[50px]" />
                   {friend.name}
                 </span>
                 <Button
@@ -127,7 +137,6 @@ const AddressBookModal: React.FC<{
           onClose={() => setRequestModalOpen(false)}
           title="친구 요청"
           explanation="친구 요청에 수락하시겠습니까?"
-          userId={selectedFriendId}
         />
       )}
       <FriendAddModal
